refactoring: documents and providers
