+++ src/utils.ts
@@ -1,0 +1,7 @@
+import { workspace, Uri } from 'vscode';
+
+export function findFiles(includeExt: string[]): Thenable<Uri[]> {
+    const searchPattern = `**/src/**/*.{${includeExt.reduce((i, j) => i + ',' + j)}}`;
+    const excludePattern = `**/node_modules/**`;
+    return workspace.findFiles(searchPattern, excludePattern);
+}
